+++ src/canceler.js
@@ -1,0 +1,14 @@
+export function Canceler () {
+  
+  if (!(this instanceof Canceler)) return new Canceler()
+
+  this.promise = new Promise(function cancelExecutor (resolve) {
+    this.resolve = resolve
+  }.bind(this))
+
+}
+
+
+Canceler.prototype.cancel = function (cancelSuccessCallback) {
+  this.resolve(cancelSuccessCallback)
+}